+++ app/controllers/concerns/dorsale/flyboy/tasks_summary_concern.rb
@@ -2,8 +2,7 @@ module Dorsale::Flyboy::TasksSummaryConcern
   extend ActiveSupport::Concern
 
   def setup_tasks_summary
-    tasks = policy_scope(::Dorsale::Flyboy::Task)
-      .where("(owner_id IS NULL and owner_type IS NULL) OR (owner_id = ? and owner_type = ?)", current_user.id, current_user.class)
+    tasks = policy_scope(::Dorsale::Flyboy::Task).where(owner: [current_user, nil])
 
     @delayed_tasks        = tasks.delayed
     @today_tasks          = tasks.today