+++ spec/dummy/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20170418070730) do
+ActiveRecord::Schema.define(version: 20170427083541) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -55,7 +55,6 @@ ActiveRecord::Schema.define(version: 20170418070730) do
     t.string   "attachable_type"
     t.string   "file"
     t.integer  "sender_id"
-    t.string   "sender_type"
     t.string   "name"
     t.datetime "created_at"
     t.datetime "updated_at"
@@ -64,7 +63,6 @@ ActiveRecord::Schema.define(version: 20170418070730) do
     t.index ["attachable_type"], name: "index_dorsale_alexandrie_attachments_on_attachable_type", using: :btree
     t.index ["attachment_type_id"], name: "index_dorsale_alexandrie_attachments_on_attachment_type_id", using: :btree
     t.index ["sender_id"], name: "index_dorsale_alexandrie_attachments_on_sender_id", using: :btree
-    t.index ["sender_type"], name: "index_dorsale_alexandrie_attachments_on_sender_type", using: :btree
   end
 
   create_table "dorsale_billing_machine_id_cards", force: :cascade do |t|
@@ -196,11 +194,9 @@ ActiveRecord::Schema.define(version: 20170418070730) do
     t.text     "text"
     t.datetime "created_at",       null: false
     t.datetime "updated_at",       null: false
-    t.string   "author_type"
     t.date     "date"
     t.string   "title"
     t.index ["author_id"], name: "index_dorsale_comments_on_author_id", using: :btree
-    t.index ["author_type"], name: "index_dorsale_comments_on_author_type", using: :btree
     t.index ["commentable_id"], name: "index_dorsale_comments_on_commentable_id", using: :btree
     t.index ["commentable_type"], name: "index_dorsale_comments_on_commentable_type", using: :btree
     t.index ["user_type"], name: "index_dorsale_comments_on_user_type", using: :btree
@@ -323,10 +319,8 @@ ActiveRecord::Schema.define(version: 20170418070730) do
     t.integer  "progress"
     t.datetime "created_at",  null: false
     t.datetime "updated_at",  null: false
-    t.string   "author_type"
     t.integer  "author_id"
     t.index ["author_id"], name: "index_dorsale_flyboy_task_comments_on_author_id", using: :btree
-    t.index ["author_type"], name: "index_dorsale_flyboy_task_comments_on_author_type", using: :btree
     t.index ["task_id"], name: "index_dorsale_flyboy_task_comments_on_task_id", using: :btree
   end
 
@@ -341,13 +335,11 @@ ActiveRecord::Schema.define(version: 20170418070730) do
     t.date     "reminder_date"
     t.datetime "created_at",        null: false
     t.datetime "updated_at",        null: false
-    t.string   "owner_type"
     t.integer  "owner_id"
     t.string   "reminder_type"
     t.integer  "reminder_duration"
     t.string   "reminder_unit"
     t.index ["owner_id"], name: "index_dorsale_flyboy_tasks_on_owner_id", using: :btree
-    t.index ["owner_type"], name: "index_dorsale_flyboy_tasks_on_owner_type", using: :btree
     t.index ["taskable_id"], name: "index_dorsale_flyboy_tasks_on_taskable_id", using: :btree
     t.index ["taskable_type"], name: "index_dorsale_flyboy_tasks_on_taskable_type", using: :btree
   end